+++ radix-engine/src/engine/wasm_runtime.rs
@@ -1,7 +1,9 @@
 use crate::engine::RuntimeError;
 use crate::engine::{HeapRENode, SystemApi};
 use crate::fee::*;
-use crate::model::{Component, ComponentInfo, ComponentState, HeapKeyValueStore, InvokeError};
+use crate::model::{
+    Component, ComponentInfoSubstate, ComponentStateSubstate, HeapKeyValueStore, InvokeError,
+};
 use crate::types::*;
 use crate::wasm::*;
 
@@ -96,8 +98,8 @@ where
 
                 // Create component
                 HeapRENode::Component(Component {
-                    info: ComponentInfo::new(package_address, blueprint_name, Vec::new()),
-                    state: ComponentState::new(state),
+                    info: ComponentInfoSubstate::new(package_address, blueprint_name, Vec::new()),
+                    state: ComponentStateSubstate::new(state),
                 })
             }
             ScryptoRENode::KeyValueStore => HeapRENode::KeyValueStore(HeapKeyValueStore::new()),
